# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion de boutique",
        "description": "API REST pour la gestion des catégories, produits, acheteurs et achats — CCP 2026",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Serveur local de développement"
        }
    ],
    "paths": {
        "/acheteurs": {
            "get": {
                "tags": [
                    "Acheteurs"
                ],
                "summary": "Lister tous les acheteurs",
                "operationId": "3048d16a74d411acda3cbc100611d321",
                "responses": {
                    "200": {
                        "description": "Liste des acheteurs"
                    }
                }
            },
            "post": {
                "tags": [
                    "Acheteurs"
                ],
                "summary": "Créer un acheteur",
                "operationId": "f6a1f76165e42643d45d60f141ce6fa6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "email"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Moussa Sarr"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "moussa.sarr@example.com"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "781234567"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Acheteur créé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/acheteurs/{id}": {
            "get": {
                "tags": [
                    "Acheteurs"
                ],
                "summary": "Afficher le détail d'un acheteur",
                "operationId": "09860c055857cc04361e86f7c71d5d68",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail de l'acheteur avec historique d'achats"
                    },
                    "404": {
                        "description": "Acheteur non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Acheteurs"
                ],
                "summary": "Modifier un acheteur",
                "operationId": "7711ae9ab9c69a6b4b86d82d763f187c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Moussa Sarr"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "moussa.sarr@example.com"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "781234567"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Acheteur modifié"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Acheteurs"
                ],
                "summary": "Supprimer un acheteur",
                "operationId": "9045802096d4fba8184cb073107e0eee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Acheteur supprimé"
                    }
                }
            }
        },
        "/acheteurs/{id}/acheter": {
            "post": {
                "tags": [
                    "Acheteurs"
                ],
                "summary": "Enregistrer un achat pour cet acheteur",
                "operationId": "01c97ac66fc872098ba823c0243905d3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "produit_id",
                                    "quantite",
                                    "date_achat"
                                ],
                                "properties": {
                                    "produit_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantite": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "date_achat": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-28"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Achat enregistré"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/categories": {
            "get": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Lister toutes les catégories",
                "operationId": "fd346ca2a3db92af72505c9cf327969b",
                "responses": {
                    "200": {
                        "description": "Liste des catégories"
                    }
                }
            },
            "post": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Créer une catégorie",
                "operationId": "4ce53e714eb24fbdd4acb9706dbe7864",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Vêtements"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Habits et accessoires"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Catégorie créée"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/categories/{id}": {
            "get": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Afficher le détail d'une catégorie",
                "operationId": "b6acb3f1c6a43306ab6fb1c408542a8d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail de la catégorie"
                    },
                    "404": {
                        "description": "Catégorie non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Modifier une catégorie",
                "operationId": "a9374f00e39c11b554a16e4f1c0c6a94",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Vêtements"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Habits et accessoires"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Catégorie modifiée"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Supprimer une catégorie",
                "operationId": "973cbeed53e164331725f7ceec9b7287",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Catégorie supprimée"
                    }
                }
            }
        },
        "/produits": {
            "get": {
                "tags": [
                    "Produits"
                ],
                "summary": "Lister tous les produits",
                "operationId": "8b52d87531fe38cafe96328a8f1dd064",
                "responses": {
                    "200": {
                        "description": "Liste des produits"
                    }
                }
            },
            "post": {
                "tags": [
                    "Produits"
                ],
                "summary": "Créer un produit",
                "operationId": "334cf1dab5c82124f1c45a4555ed7795",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prix",
                                    "stock",
                                    "categorie_id"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Écouteurs Bluetooth"
                                    },
                                    "prix": {
                                        "type": "number",
                                        "example": 25000
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Sans fil, autonomie 8h"
                                    },
                                    "categorie_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produit créé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/produits/{id}": {
            "get": {
                "tags": [
                    "Produits"
                ],
                "summary": "Afficher le détail d'un produit",
                "operationId": "9c5d60584961429a9c2a243463183cec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail du produit"
                    },
                    "404": {
                        "description": "Produit non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Produits"
                ],
                "summary": "Modifier un produit",
                "operationId": "42c2896c9129639cbcb0094ebb631890",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Écouteurs Bluetooth"
                                    },
                                    "prix": {
                                        "type": "number",
                                        "example": 25000
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Sans fil, autonomie 8h"
                                    },
                                    "categorie_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produit modifié"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Produits"
                ],
                "summary": "Supprimer un produit",
                "operationId": "f9e972f1ce50047b09fd6db1ee1ec82d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Produit supprimé"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Acheteurs",
            "description": "Acheteurs"
        },
        {
            "name": "Catégories",
            "description": "Catégories"
        },
        {
            "name": "Produits",
            "description": "Produits"
        }
    ]
}